wiring up props
